web/requirements.txt
fastapi
uvicorn
python-dotenv
pillow==9.5.0
SQLAlchemy
asyncpg
alembic
psycopg2
python-multipart
imageio[ffmpeg]